resolved: fixed issue with secret not getting deciphered
The issue was due to how the secret I was attempting to access was encoded within my database. I had directly stored a base64 encoded secret value in mongodb. After switching to a method of storing the base64 value as string in the database, this made the decryption process a lot more streamlined.